% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rxes.R
\name{createdf}
\alias{createdf}
\title{Supporting function - transform map of traces into dataframe}
\usage{
createdf(res)
}
\arguments{
\item{res}{javaref object containing map of flatXTraces}
}
\value{
data.frame with traces
}
\description{
Supporting function - transform map of traces into dataframe
}

