% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{processfilter}
\alias{processfilter}
\title{Create filter for xestool calls based on provided parameters}
\usage{
processfilter(resources, groups, roles, eventcount, tracestart, traceend,
  eventnames, tracestartwday, traceendwday, transitions)
}
\arguments{
\item{resources}{list of resources present in trace's event to pass filter. MULTI for multiple resources per trace}

\item{groups}{list of resources present in trace's event to pass filter. MULTI for multiple groups per trace}

\item{roles}{list of resources present in trace's event to pass filter. MULTI for multiple roles per trace}

\item{eventcount}{range of events per trace to pass filter}

\item{tracestart}{range of trace start dates to pass filter}

\item{traceend}{range of trace end dates to pass filter}

\item{eventnames}{list of events trace should contain (at least one) to pass filter}

\item{tracestartwday}{list of trace start DoW (at least one)  to pass filter}

\item{traceendwday}{list of trace end DoW (at least one) to pass filter}

\item{transitions}{list of event transitions (at least one) to pass filter}
}
\value{
filter object (hashmap)
}
\description{
Create filter for xestool calls based on provided parameters
}

