% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rxes.R
\name{xes.getEventList}
\alias{xes.getEventList}
\title{Retreive event list form traces matching filter}
\usage{
xes.getEventList(xes, ...)
}
\arguments{
\item{xes}{reference to xes class object created via \code{\link{xes.init}}}

\item{...}{parameters to be passed to filter creation function \code{\link{xes.xes.processfilter}}}
}
\value{
list of events
}
\description{
Retreive event list form traces matching filter
}

