% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rxes.R
\name{xes.getEventShare}
\alias{xes.getEventShare}
\title{Calculate average share of each event class duration of total trace duration over full log}
\usage{
xes.getEventShare(xes, median = FALSE)
}
\arguments{
\item{xes}{reference to xes class object created via \code{\link{xes.init}}}

\item{median}{if true, use median instead of mean}
}
\value{
data.frame with event shares
}
\description{
Calculate average share of each event class duration of total trace duration over full log
}

