% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rxes.R
\name{xes.getFullSubTraceList}
\alias{xes.getFullSubTraceList}
\title{Return list of all trace segments limited by event names with attributes}
\usage{
xes.getFullSubTraceList(xesTools, startEvent, endEvent)
}
\arguments{
\item{xesTools}{reference to XEStools class object created via \code{\link{xes.init}}}

\item{startEvent}{name of first event in trace segment}

\item{endEvent}{name of last event in trace segment}
}
\value{
data.frame with traces
}
\description{
Return list of all trace segments limited by event names with attributes
}

