% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxes.R
\name{xes.getTraceDurations}
\alias{xes.getTraceDurations}
\title{Return data.frame with the trace name (concept:name) and trace duration expressed in seconds}
\usage{
xes.getTraceDurations(xesTools)
}
\arguments{
\item{xesTools}{reference to XEStools class object created via \code{\link{xes.init}}}
}
\value{
data.frame with trace durations
}
\description{
Return data.frame with the trace name (concept:name) and trace duration expressed in seconds
}

