% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rxes.R
\name{xes.init}
\alias{xes.init}
\title{Initialize XES Log processing class}
\usage{
xes.init()
}
\value{
reference to xes object
}
\description{
\code{xes.init} returns pointer to the javaClass with empty log
}
\details{
You should pass this reference to the other funcions, so we can use cache and avoiud log reload on any
operation
}

