% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rxes.R
\name{xes.parseLog}
\alias{xes.parseLog}
\title{Loads and parses XES log in GZIP xml format}
\usage{
xes.parseLog(xesTools, filename)
}
\arguments{
\item{xesTools}{reference to XEStools class object created via \code{\link{xes.init}}}

\item{filename}{string pointing to the event log file}
}
\value{
boolean success flag
}
\description{
loads and saves in the class object log specified by the filename
}

