% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{xes.processfilter}
\alias{xes.processfilter}
\title{Create filter for xestool calls based on provided parameters}
\usage{
xes.processfilter(resources = NULL, groups = NULL, roles = NULL,
  eventcount = NULL, tracestart = NULL, traceend = NULL,
  eventnames = NULL, tracestartwday = NULL, traceendwday = NULL,
  transitions = NULL, tracenames = NULL)
}
\arguments{
\item{resources}{list of resources present in trace's event to pass filter. MULTI for multiple resources per trace}

\item{groups}{list of resources present in trace's event to pass filter. MULTI for multiple groups per trace}

\item{roles}{list of resources present in trace's event to pass filter. MULTI for multiple roles per trace}

\item{eventcount}{range of events per trace to pass filter}

\item{tracestart}{range of trace start dates to pass filter}

\item{traceend}{range of trace end dates to pass filter}

\item{eventnames}{list of events trace should contain (at least one) to pass filter}

\item{tracestartwday}{list of trace start DoW (at least one)  to pass filter}

\item{traceendwday}{list of trace end DoW (at least one)  to pass filter#'}

\item{transitions}{list of event transitions statuses to pass filter}

\item{tracenames}{list of trace names to pass filter}
}
\value{
filter object (hashmap)
}
\description{
Create filter for xestool calls based on provided parameters
}

