% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxes.R
\name{xes.traceCount}
\alias{xes.traceCount}
\title{Return count of traces in the event log}
\usage{
xes.traceCount(xesTools)
}
\arguments{
\item{xesTools}{reference to XEStools class object created via \code{\link{xes.init}}}
}
\value{
success flag
}
\description{
Return count of traces in the event log
}

