\name{rxes-package}
\alias{rxes-package}
\alias{rxes}
\docType{package}
\title{
\packageTitle{rxes}
}
\description{
\packageDescription{rxes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rxes}
\packageIndices{rxes}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rxes}

Maintainer: \packageMaintainer{rxes}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
