% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rxes.R
\name{xes.getFullTraceList}
\alias{xes.getFullTraceList}
\title{Return list of all traces with attributes}
\usage{
xes.getFullTraceList(xes, ...)
}
\arguments{
\item{xes}{reference to xes class object created via \code{\link{xes.init}}}
}
\value{
data.frame with traces
}
\description{
Return list of all traces with attributes
}

