% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rxes.R
\name{xes.getFullTraceList}
\alias{xes.getFullTraceList}
\title{Return list of all traces with attributes}
\usage{
xes.getFullTraceList(xes, resources = NULL, groups = NULL, roles = NULL,
  eventcount = NULL, tracestart = NULL, traceend = NULL,
  eventnames = NULL, tracestartwday = NULL, traceendwday = NULL,
  transitions = NULL)
}
\arguments{
\item{xes}{reference to xes class object created via \code{\link{xes.init}}}

\item{resources}{list of resources present in trace's event to pass filter. MULTI for multiple resources per trace}

\item{groups}{list of resources present in trace's event to pass filter. MULTI for multiple groups per trace}

\item{roles}{list of resources present in trace's event to pass filter. MULTI for multiple roles per trace}

\item{eventcount}{range of events per trace to pass filter}

\item{tracestart}{range of trace start dates to pass filter}

\item{eventnames}{list of events trace should contain (at least one) to pass filter}

\item{tracestartwday}{list of trace start DoW (at least one)  to pass filter}

\item{tracesend}{range of trace end dates to pass filter}

\item{tracesendwday}{list of trace end DoW (at least one) to pass filter}

\item{tracesend}{list of event transitions (at least one) to pass filter}
}
\value{
data.frame with traces
}
\description{
Return list of all traces with attributes
}

