% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rxes.R
\name{xes.getTraceDurations}
\alias{xes.getTraceDurations}
\title{Return data.frame with the trace name (concept:name) and trace duration expressed in seconds}
\usage{
xes.getTraceDurations(xes)
}
\arguments{
\item{xes}{reference to xes class object created via \code{\link{xes.init}}}
}
\value{
data.frame with trace durations
}
\description{
Return data.frame with the trace name (concept:name) and trace duration expressed in seconds
}

